Thermoelectric Manipulator Fix
The thermoelectric manipulator had a problem where it could not make fluids evaporate.
Looking at the code I found that the issue is that it wasn't findng the evaporated gas in the mapped gasses hashmap. I changed the index and it now returns the evaporated gas, making the check to work properly and the manipulator to work